telegram_bot/requirements.txt
fastapi
uvicorn[standard]
aiogram
python-dotenv
requests
httpx